decrypt_and_verify -> verify_and_decrypt
We are using MessageEncryptor as the basis of an AEAD scheme which encrypts
first, then MACs the ciphertext. Therefore the decrypt operation should first
verify the MAC prior to decrypting.

This actually was the previous behavior, but for whatever reason the upstream
ActiveSupport code chose the name "decrypt_and_verify" despite this not being
the ordering it performs the operation in.

We will use a more correct name in Cryptor.